+++ src/components/SearchBar/searchBarChangelogData.ts
@@ -10,6 +10,13 @@ import { ChangelogData } from "../../utils/ComponentChangelogTable";
 
 export const changelogData: ChangelogData[] = [
   {
+    date: "Prerelease",
+    version: "Prerelease",
+    type: "Update",
+    affects: ["Styles"],
+    notes: ["Updates Searchbar style to remove stacking on mobile breakpoints."],
+  },
+  {
     date: "2024-06-20",
     version: "3.1.6",
     type: "Update",