+++ CHANGELOG.md
@@ -11,6 +11,7 @@ Currently, this repo is in Prerelease. When it is released, this project will ad
 ### Updates
 
 - Updates Banner component to allow for HTML content in the `content` prop when passed as a string.
+- Updates `Searchbar` component styles on mobile breakpoints
 
 ## 3.2.0 (July 25, 2024)